--- backend/src/todo/todo.resolver.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { TodoResolver } from './todo.resolver';
-
-describe('TodoResolver', () => {
-  let resolver: TodoResolver;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [TodoResolver],
-    }).compile();
-
-    resolver = module.get<TodoResolver>(TodoResolver);
-  });
-
-  it('should be defined', () => {
-    expect(resolver).toBeDefined();
-  });
-});
