+++ backend/tsconfig.json
@@ -1,7 +1,10 @@
 {
   "compilerOptions": {
     "module": "commonjs",
-    "types": ["@types/jest", "node"],
+    "types": [
+      "@types/jest",
+      "node"
+    ],
     "declaration": true,
     "removeComments": true,
     "emitDecoratorMetadata": true,
@@ -17,6 +20,7 @@
     "forceConsistentCasingInFileNames": true,
     "noImplicitAny": false,
     "strictBindCallApply": false,
-    "noFallthroughCasesInSwitch": false
+    "noFallthroughCasesInSwitch": false,
+    "moduleResolution": "node"
   }
 }
